Extend Control interface
With this commit it is easier to add new controls, even controls
not defined in the gopkg.in/ldap.vX package. The new Control just
needs to implement the interface, no knowledge of the control
has to be in the gopkg.in/ldap.vX package.

Note: current state is that both VChu Password controls do not
know how to Encode() or Describe() themself, this should be done
in a separate commit.

Any Control needs to call RegisterControl(oid, description, Control)
to make it self known to the package like the Manage DSA IT control:
```go
func init() {
   RegisterControl(ControlTypeManageDsaIT, "Manage DSA IT", ControlManageDsaIT(false))
}
```

Changes:
* extend Control interface{}:
** add `Decode(bool, *ber.Packet) (Control, error)` - this
   decodes the Control. To support non pointer controls, we
   need to return the control...
** add `Describe(*ber.Packet)` to add descritptions to the control value
** add missing functions to the existing Controls
** convert ControlManageDsaIT to `type ControlManageDsaIT bool` to show
   it's working

* split all existing Controls (except for ControlString)
  into own file

* DecodeControl():
** returns Control and error (API change)
** just extracts OID, Criticality and the control value,
   calls GetControl(OID) and let the newly initialized Control
   Decode() the value and set Criticality

* utilities:
** add `RegisterControl(oid, description string, ctrl Control)`
   to add a Control to the known controls, call RegisterControl()
   for each existing Control via the init() function
** add `GetControl(oid) Control` to get a newly initialized
   Control for the OID oid
** add `ControlDescription(oid) string` to get the description
   for the Control

Maybe change:
* Return error when passed Control is nil in RegisterControl?
* rename ControlDescription to GetControlDescription?
